moved ThreadedInterface to dedicated header
